Fixes and refactor for custom pretraining loop (#12319)
