+++ nemo/tron/state.py
@@ -25,9 +25,9 @@ from megatron.core.utils import StragglerDetector
 from torch.distributed.checkpoint.stateful import Stateful
 
 from nemo.tron.config import ConfigContainer
-from nemo.tron.dist_signal_handler import DistributedSignalHandler
 from nemo.tron.tokenizers.tokenizer import build_tokenizer
 from nemo.tron.utils.common_utils import dump_dataclass_to_yaml, get_rank_safe, get_world_size_safe
+from nemo.tron.utils.sig_utils import DistributedSignalHandler
 
 
 def _timers_write_to_wandb(
