+++ nemo/tron/utils/theoretical_memory_utils.py
